Move top-level `notion-core` modules in subdirectories in preparation for moving their serialization modules into submodules of those directories, and move `serial::{catalog, index}` into `catalog::serial`.
