+++ crates/notion-core/src/manifest/mod.rs
