+++ crates/notion-core/src/serial/mod.rs
@@ -1,8 +1,6 @@
 //! Provides utilities for serializing and deserializing file formats.
 
-pub mod catalog;
 pub mod config;
-pub mod index;
 pub mod manifest;
 pub mod plugin;
 pub mod version;
